refactor: cleanup contracts folder
