+++ .gitignore
@@ -154,3 +154,5 @@ packages/cli/contractAddresses.old.json
 packages/circuits/circom/test
 
 **/ts/__benchmarks__/results/**
+
+proofs
