+++ packages/contracts/contracts/gatekeepers/interfaces/IHats.sol
