+++ packages/contracts/contracts/gatekeepers/interfaces/ISemaphore.sol
