--- packages/contracts/contracts/trees/zeros/MerkleBinary0.sol
@@ -1,43 +1,0 @@
-// SPDX-License-Identifier: MIT
-pragma solidity ^0.8.20;
-
-abstract contract MerkleZeros {
-  uint256[33] internal zeros;
-
-  // Binary tree zeros (0)
-  constructor() {
-    zeros[0] = uint256(0);
-    zeros[1] = uint256(14744269619966411208579211824598458697587494354926760081771325075741142829156);
-    zeros[2] = uint256(7423237065226347324353380772367382631490014989348495481811164164159255474657);
-    zeros[3] = uint256(11286972368698509976183087595462810875513684078608517520839298933882497716792);
-    zeros[4] = uint256(3607627140608796879659380071776844901612302623152076817094415224584923813162);
-    zeros[5] = uint256(19712377064642672829441595136074946683621277828620209496774504837737984048981);
-    zeros[6] = uint256(20775607673010627194014556968476266066927294572720319469184847051418138353016);
-    zeros[7] = uint256(3396914609616007258851405644437304192397291162432396347162513310381425243293);
-    zeros[8] = uint256(21551820661461729022865262380882070649935529853313286572328683688269863701601);
-    zeros[9] = uint256(6573136701248752079028194407151022595060682063033565181951145966236778420039);
-    zeros[10] = uint256(12413880268183407374852357075976609371175688755676981206018884971008854919922);
-    zeros[11] = uint256(14271763308400718165336499097156975241954733520325982997864342600795471836726);
-    zeros[12] = uint256(20066985985293572387227381049700832219069292839614107140851619262827735677018);
-    zeros[13] = uint256(9394776414966240069580838672673694685292165040808226440647796406499139370960);
-    zeros[14] = uint256(11331146992410411304059858900317123658895005918277453009197229807340014528524);
-    zeros[15] = uint256(15819538789928229930262697811477882737253464456578333862691129291651619515538);
-    zeros[16] = uint256(19217088683336594659449020493828377907203207941212636669271704950158751593251);
-    zeros[17] = uint256(21035245323335827719745544373081896983162834604456827698288649288827293579666);
-    zeros[18] = uint256(6939770416153240137322503476966641397417391950902474480970945462551409848591);
-    zeros[19] = uint256(10941962436777715901943463195175331263348098796018438960955633645115732864202);
-    zeros[20] = uint256(15019797232609675441998260052101280400536945603062888308240081994073687793470);
-    zeros[21] = uint256(11702828337982203149177882813338547876343922920234831094975924378932809409969);
-    zeros[22] = uint256(11217067736778784455593535811108456786943573747466706329920902520905755780395);
-    zeros[23] = uint256(16072238744996205792852194127671441602062027943016727953216607508365787157389);
-    zeros[24] = uint256(17681057402012993898104192736393849603097507831571622013521167331642182653248);
-    zeros[25] = uint256(21694045479371014653083846597424257852691458318143380497809004364947786214945);
-    zeros[26] = uint256(8163447297445169709687354538480474434591144168767135863541048304198280615192);
-    zeros[27] = uint256(14081762237856300239452543304351251708585712948734528663957353575674639038357);
-    zeros[28] = uint256(16619959921569409661790279042024627172199214148318086837362003702249041851090);
-    zeros[29] = uint256(7022159125197495734384997711896547675021391130223237843255817587255104160365);
-    zeros[30] = uint256(4114686047564160449611603615418567457008101555090703535405891656262658644463);
-    zeros[31] = uint256(12549363297364877722388257367377629555213421373705596078299904496781819142130);
-    zeros[32] = uint256(21443572485391568159800782191812935835534334817699172242223315142338162256601);
-  }
-}
