--- packages/contracts/contracts/trees/zeros/MerkleBinaryBlankSl.sol
@@ -1,43 +1,0 @@
-// SPDX-License-Identifier: MIT
-pragma solidity ^0.8.20;
-
-abstract contract MerkleZeros {
-  uint256[33] internal zeros;
-
-  // Binary tree zeros (hash of a blank state leaf)
-  constructor() {
-    zeros[0] = uint256(6769006970205099520508948723718471724660867171122235270773600567925038008762);
-    zeros[1] = uint256(2972820301952105722688860985556183033855705951263221082702981787813754939537);
-    zeros[2] = uint256(19009473369953096352828532459942637819279786575057870804186038131433538383332);
-    zeros[3] = uint256(1877001762518233819645599208989578372605193385355680834239714249281096297174);
-    zeros[4] = uint256(4022598852800694816938652741439614774645858989706174527409714109784047480520);
-    zeros[5] = uint256(8078617093048295855521451309865989496051030103472138252021705658681696298712);
-    zeros[6] = uint256(21861637049723057871988413507302821095913894718242489848472531680353400271584);
-    zeros[7] = uint256(2969626195902860050407584814596940245443093107470116547781577350415736914038);
-    zeros[8] = uint256(13863086449569754493134198846069090996475357615094865751949144794620598051673);
-    zeros[9] = uint256(13774233155966252113965527228795435224641075024674384267997743867571011718458);
-    zeros[10] = uint256(7674682532432601125535053858292577379388329393276537570517515727197672122157);
-    zeros[11] = uint256(2657471847139856346360223652201172662911313292042510535836997980857168085414);
-    zeros[12] = uint256(14112562742724116016492623819773686970029672095023612838615540190985426106768);
-    zeros[13] = uint256(16966520284141749853106006448832965932249937855809150844697400390499975107456);
-    zeros[14] = uint256(21146121663662200258116396149536742745305242191891337170899444969488030502620);
-    zeros[15] = uint256(8395571901509192935479743034608666551563743095742598750914087478677907730358);
-    zeros[16] = uint256(11584898446168752024843587018551921614604785083342073076015560385003528300499);
-    zeros[17] = uint256(19681365563800708744156562671961079617734353445922751560400662591064339349816);
-    zeros[18] = uint256(11060693795061987995391612467169498625108376769265861980249917517984263067473);
-    zeros[19] = uint256(20136055137568042031318427040358591430196153124171666293804511634641041409480);
-    zeros[20] = uint256(10438448879123510479428288344427042332522761183009746406441238260861529360499);
-    zeros[21] = uint256(20302411580043873005239406811066876697276902025885155920151067303221158887923);
-    zeros[22] = uint256(16905699456770804689394621400052823445587122726651394178036372609288266146575);
-    zeros[23] = uint256(13317924909658910751179983108234689413063120680580702936091220805509299490708);
-    zeros[24] = uint256(11624463897690689883938167321830091369950933831231839575225419984927228390345);
-    zeros[25] = uint256(12388077003631746290497429926117583834311703848735670874049584990731919769551);
-    zeros[26] = uint256(16641943593086083573943184041147806885253724243247212515325749241831788827569);
-    zeros[27] = uint256(8675770901378242337954792996483564563211065498082968464791979179678744114204);
-    zeros[28] = uint256(3741944068643598116715410464277276913339851849923986024648161859457213369743);
-    zeros[29] = uint256(9365051374992868354747065577256691008852056444829383197903446097138255771103);
-    zeros[30] = uint256(19608043542461863702809013760105552654336523908709289008189330402608282498922);
-    zeros[31] = uint256(15116478429455923389320892447700153271977917184085737305957533984219061034768);
-    zeros[32] = uint256(13372161856163346716845871420623647679532631520878788090782842562075678687737);
-  }
-}
