--- packages/contracts/contracts/trees/zeros/MerkleQuinary0.sol
@@ -1,43 +1,0 @@
-// SPDX-License-Identifier: MIT
-pragma solidity ^0.8.20;
-
-abstract contract MerkleZeros {
-  uint256[33] internal zeros;
-
-  // Quinary tree zeros (0)
-  constructor() {
-    zeros[0] = uint256(0);
-    zeros[1] = uint256(14655542659562014735865511769057053982292279840403315552050801315682099828156);
-    zeros[2] = uint256(19261153649140605024552417994922546473530072875902678653210025980873274131905);
-    zeros[3] = uint256(21526503558325068664033192388586640128492121680588893182274749683522508994597);
-    zeros[4] = uint256(20017764101928005973906869479218555869286328459998999367935018992260318153770);
-    zeros[5] = uint256(16998355316577652097112514691750893516081130026395813155204269482715045879598);
-    zeros[6] = uint256(2612442706402737973181840577010736087708621987282725873936541279764292204086);
-    zeros[7] = uint256(17716535433480122581515618850811568065658392066947958324371350481921422579201);
-    zeros[8] = uint256(17437916409890180001398333108882255895598851862997171508841759030332444017770);
-    zeros[9] = uint256(20806704410832383274034364623685369279680495689837539882650535326035351322472);
-    zeros[10] = uint256(6821382292698461711184253213986441870942786410912797736722948342942530789476);
-    zeros[11] = uint256(5916648769022832355861175588931687601652727028178402815013820610204855544893);
-    zeros[12] = uint256(8979092375429814404031883906996857902016801693563521316025319397481362525766);
-    zeros[13] = uint256(2921214989930864339537708350754648834701757280474461132621735242274490553963);
-    zeros[14] = uint256(8930183771974746972686153669144011224662017420905079900118160414492327314176);
-    zeros[15] = uint256(235368305313252659057202520253547068193638476511860624369389264358598810396);
-    zeros[16] = uint256(11594802086624841314469980089838552727386894436467147447204224403068085066609);
-    zeros[17] = uint256(6527402365840056202903190531155009847198979121365335038364206235405082926579);
-    zeros[18] = uint256(7890267294950363768070024023123773394579161137981585347919627664365669195485);
-    zeros[19] = uint256(7743021844925994795008658518659888250339967931662466893787320922384170613250);
-    zeros[20] = uint256(3315762791558236426429898223445373782079540514426385620818139644150484427120);
-    zeros[21] = uint256(12047412166753578299610528762227103229354276396579409944098869901020016693788);
-    zeros[22] = uint256(7346375653460369101190037700418084792046605818930533590372465301789036536);
-    zeros[23] = uint256(16686328169837855831280640081580124364395471639440725186157725609010405016551);
-    zeros[24] = uint256(19105160640579355001844872723857900201603625359252284777965070378555675817865);
-    zeros[25] = uint256(17054399483511247964029303840879817843788388567881464290309597953132679359256);
-    zeros[26] = uint256(5296258093842160235704190490839277292290579093574356735268980000023915581697);
-    zeros[27] = uint256(8993437003863084469472897416707962588904917898547964184966432920162387360131);
-    zeros[28] = uint256(7234267096117283161039619077058835089667467648437312224957703988301725566335);
-    zeros[29] = uint256(21640448288319814375882234036901598260365718394023649234526744669922384765526);
-    zeros[30] = uint256(1595567811792178436811872247033324109773732075641399161664435302467654689847);
-    zeros[31] = uint256(15291095622175285816966181294098521638815701170497062413595539727181544870101);
-    zeros[32] = uint256(15837036953038489303182430773663047564827202645548797032627170282475341436016);
-  }
-}
