+++ packages/contracts/package.json
@@ -47,19 +47,17 @@
     "test:tallyNonQv": "pnpm run test ./tests/TallyNonQv.test.ts",
     "test:hasher": "pnpm run test ./tests/Hasher.test.ts",
     "test:utilities": "pnpm run test ./tests/Utilities.test.ts",
-    "test:signupGatekeeper": "pnpm run test ./tests/SignUpGatekeeper.test.ts",
+    "test:signupGatekeeper": "pnpm run test ./tests/gatekeepers/SignUpGatekeeper.test.ts",
     "test:verifier": "pnpm run test ./tests/Verifier.test.ts",
     "test:hasherBenchmarks": "pnpm run test ./tests/HasherBenchmarks.test.ts",
     "test:vkRegistry": "pnpm run test ./tests/VkRegistry.test.ts",
     "test:pollFactory": "pnpm run test ./tests/PollFactory.test.ts",
-    "test:eas_gatekeeper": "pnpm run test ./tests/EASGatekeeper.test.ts",
-    "test:hats_gatekeeper": "pnpm run test ./tests/HatsGatekeeper.test.ts",
-    "test:gitcoin_gatekeeper": "pnpm run test ./tests/GitcoinPassportGatekeeper.test.ts",
-    "test:zupass_gatekeeper": "pnpm run test ./tests/ZupassGatekeeper.test.ts",
-    "test:semaphore_gatekeeper": "pnpm run test ./tests/SemaphoreGatekeeper.test.ts",
-    "test:anon_aadhaar_gatekeeper": "pnpm run test ./tests/AnonAadhaarGatekeeper.test.ts",
-    "test:simpleProjectRegistry": "pnpm run test ./tests/extensions/SimpleProjectRegistry.test.ts",
-    "test:simplePayout": "pnpm run test ./tests/extensions/SimplePayout.test.ts",
+    "test:eas_gatekeeper": "pnpm run test ./tests/gatekeepers/EASGatekeeper.test.ts",
+    "test:hats_gatekeeper": "pnpm run test ./tests/gatekeepers/HatsGatekeeper.test.ts",
+    "test:gitcoin_gatekeeper": "pnpm run test ./tests/gatekeepers/GitcoinPassportGatekeeper.test.ts",
+    "test:zupass_gatekeeper": "pnpm run test ./tests/gatekeepers/ZupassGatekeeper.test.ts",
+    "test:semaphore_gatekeeper": "pnpm run test ./tests/gatekeepers/SemaphoreGatekeeper.test.ts",
+    "test:anon_aadhaar_gatekeeper": "pnpm run test ./tests/gatekeepers/AnonAadhaarGatekeeper.test.ts",
     "deploy": "hardhat deploy-full",
     "deploy-poll": "hardhat deploy-poll",
     "verify": "hardhat verify-full",
